start cleaning up server
