+++ packages/paperclip-designer/src/reducers/designer.ts
@@ -14,9 +14,7 @@ import {
   mergeBoxes,
   getActiveFrameIndex,
   isExpanded,
-  getPreviewChildren,
   getCurrentPreviewFrameBoxes,
-  updateShared,
   DesignerState,
   SyncLocationMode,
   pruneDeletedNodes,
