+++ packages/paperclip-designer/src/server/core/disposable.ts
@@ -1,0 +1,3 @@
+export type Disposable = {
+  dispose: () => void;
+};
